@@ -162,8 +162,8 @@ class StreamChannelHeader extends StatelessWidget
           showMessage: showConnectionStateTile && showStatus,
           message: statusString,
           child: AppBar(
-            toolbarTextStyle: theme.textTheme.bodyText2,
-            titleTextStyle: theme.textTheme.headline6,
+            toolbarTextStyle: theme.textTheme.bodyMedium,
+            titleTextStyle: theme.textTheme.titleLarge,
             systemOverlayStyle: theme.brightness == Brightness.dark
                 ? SystemUiOverlayStyle.light
                 : SystemUiOverlayStyle.dark,
