@@ -80,8 +80,8 @@ class StreamGalleryHeader extends StatelessWidget
     final galleryHeaderThemeData = StreamGalleryHeaderTheme.of(context);
     final theme = Theme.of(context);
     return AppBar(
-      toolbarTextStyle: theme.textTheme.bodyText2,
-      titleTextStyle: theme.textTheme.headline6,
+      toolbarTextStyle: theme.textTheme.bodyMedium,
+      titleTextStyle: theme.textTheme.titleLarge,
       systemOverlayStyle: theme.brightness == Brightness.dark
           ? SystemUiOverlayStyle.light
           : SystemUiOverlayStyle.dark,
