@@ -149,8 +149,8 @@ class StreamThreadHeader extends StatelessWidget
     final theme = Theme.of(context);
     return AppBar(
       automaticallyImplyLeading: false,
-      toolbarTextStyle: theme.textTheme.bodyText2,
-      titleTextStyle: theme.textTheme.headline6,
+      toolbarTextStyle: theme.textTheme.bodyMedium,
+      titleTextStyle: theme.textTheme.titleLarge,
       systemOverlayStyle: theme.brightness == Brightness.dark
           ? SystemUiOverlayStyle.light
           : SystemUiOverlayStyle.dark,
