@@ -409,7 +409,7 @@ class StreamChannelListErrorWidget extends StatelessWidget {
                 TextSpan(text: context.translations.loadingChannelsError),
               ],
             ),
-            style: Theme.of(context).textTheme.headline6,
+            style: Theme.of(context).textTheme.titleLarge,
           ),
           TextButton(
             onPressed: onPressed,
