@@ -38,30 +38,11 @@ ChannelModel _$ChannelModelFromJson(Map<String, dynamic> json) => ChannelModel(
       cooldown: json['cooldown'] as int? ?? 0,
     );
 
-Map<String, dynamic> _$ChannelModelToJson(ChannelModel instance) {
-  final val = <String, dynamic>{
-    'id': instance.id,
-    'type': instance.type,
-  };
-
-  void writeNotNull(String key, dynamic value) {
-    if (value != null) {
-      val[key] = value;
-    }
-  }
-
-  writeNotNull('cid', readonly(instance.cid));
-  writeNotNull('own_capabilities', readonly(instance.ownCapabilities));
-  writeNotNull('config', readonly(instance.config));
-  writeNotNull('created_by', readonly(instance.createdBy));
-  val['frozen'] = instance.frozen;
-  writeNotNull('last_message_at', readonly(instance.lastMessageAt));
-  writeNotNull('created_at', readonly(instance.createdAt));
-  writeNotNull('updated_at', readonly(instance.updatedAt));
-  writeNotNull('deleted_at', readonly(instance.deletedAt));
-  writeNotNull('member_count', readonly(instance.memberCount));
-  val['cooldown'] = instance.cooldown;
-  val['extra_data'] = instance.extraData;
-  writeNotNull('team', readonly(instance.team));
-  return val;
-}
+Map<String, dynamic> _$ChannelModelToJson(ChannelModel instance) =>
+    <String, dynamic>{
+      'id': instance.id,
+      'type': instance.type,
+      'frozen': instance.frozen,
+      'cooldown': instance.cooldown,
+      'extra_data': instance.extraData,
+    };
