@@ -20,22 +20,9 @@ Reaction _$ReactionFromJson(Map<String, dynamic> json) => Reaction(
       extraData: json['extra_data'] as Map<String, dynamic>? ?? const {},
     );
 
-Map<String, dynamic> _$ReactionToJson(Reaction instance) {
-  final val = <String, dynamic>{
-    'message_id': instance.messageId,
-    'type': instance.type,
-  };
-
-  void writeNotNull(String key, dynamic value) {
-    if (value != null) {
-      val[key] = value;
-    }
-  }
-
-  writeNotNull('created_at', readonly(instance.createdAt));
-  writeNotNull('user', readonly(instance.user));
-  val['score'] = instance.score;
-  writeNotNull('user_id', readonly(instance.userId));
-  val['extra_data'] = instance.extraData;
-  return val;
-}
+Map<String, dynamic> _$ReactionToJson(Reaction instance) => <String, dynamic>{
+      'message_id': instance.messageId,
+      'type': instance.type,
+      'score': instance.score,
+      'extra_data': instance.extraData,
+    };
