@@ -10,41 +10,13 @@ export 'package:logging/logging.dart' show Logger, Level, LogRecord;
 export 'package:rate_limiter/rate_limiter.dart';
 export 'package:uuid/uuid.dart';
 
-export './src/core/api/attachment_file_uploader.dart';
-export './src/core/api/requests.dart';
-export './src/core/api/requests.dart';
-export './src/core/api/responses.dart';
-export './src/core/api/stream_chat_api.dart' show PushProvider;
-export './src/core/error/error.dart';
-export './src/core/models/action.dart';
-export './src/core/models/attachment.dart';
-export './src/core/models/attachment_file.dart';
-export './src/core/models/channel_config.dart';
-export './src/core/models/channel_model.dart';
-export './src/core/models/channel_state.dart';
-export './src/core/models/command.dart';
-export './src/core/models/device.dart';
-export './src/core/models/event.dart';
-export './src/core/models/filter.dart' show Filter;
-export './src/core/models/member.dart';
-export './src/core/models/message.dart';
-export './src/core/models/mute.dart';
-export './src/core/models/own_user.dart';
-export './src/core/models/reaction.dart';
-export './src/core/models/read.dart';
-export './src/core/models/user.dart';
-export './src/core/util/extension.dart';
-export './src/db/chat_persistence_client.dart';
-export './src/event_type.dart';
-export './src/permission_type.dart';
-export './src/ws/connection_status.dart';
 export 'src/client/channel.dart';
 export 'src/client/client.dart';
 export 'src/client/key_stroke_handler.dart';
-export 'src/core/api/attachment_file_uploader.dart' show AttachmentFileUploader;
-export 'src/core/api/requests.dart';
+export 'src/core/api/attachment_file_uploader.dart';
 export 'src/core/api/requests.dart';
 export 'src/core/api/responses.dart';
+export 'src/core/api/stream_chat_api.dart' show PushProvider;
 export 'src/core/api/stream_chat_api.dart';
 export 'src/core/error/error.dart';
 export 'src/core/models/action.dart';
@@ -68,4 +40,5 @@ export 'src/core/platform_detector/platform_detector.dart';
 export 'src/core/util/extension.dart';
 export 'src/db/chat_persistence_client.dart';
 export 'src/event_type.dart';
+export 'src/permission_type.dart';
 export 'src/ws/connection_status.dart';
