@@ -516,6 +516,9 @@ void main() {
     });
 
     setUp(() async {
+      when(() => persistence.updateLastSyncAt(any()))
+          .thenAnswer((_) => Future.value());
+      when(persistence.getLastSyncAt).thenAnswer((_) async => null);
       client = StreamChatClient(apiKey, chatApi: api, ws: ws)
         ..chatPersistenceClient = persistence;
       await client.connectUser(user, token);
@@ -532,9 +535,12 @@ void main() {
       test(
         '''should update persistence connectionInfo and lastSync when sync succeeds''',
         () async {
+          // persistence.updateLastSyncAt might be called
+          // when connecting the user.
+          // Resetting the logs so we start counting invocations correctly.
+          reset(persistence);
           const cids = ['test-cid-1', 'test-cid-2', 'test-cid-3'];
           final lastSyncAt = DateTime.now();
-
           when(() => api.general.sync(cids, lastSyncAt))
               .thenAnswer((_) async => SyncResponse()
                 ..events = [
@@ -567,6 +573,10 @@ void main() {
       test(
         'should work fine if persistence contains sync params',
         () async {
+          // persistence.updateLastSyncAt might be called
+          // when connecting the user.
+          // Resetting the logs so we start counting invocations correctly.
+          reset(persistence);
           const cids = ['test-cid-1', 'test-cid-2', 'test-cid-3'];
           final lastSyncAt = DateTime.now();
 
