@@ -123,7 +123,9 @@ class StreamAttachmentHandler extends StreamAttachmentHandlerBase {
 
     final tempDir = await getTemporaryDirectory();
     final tempPath = Uri.file(tempDir.path, windows: CurrentPlatform.isWindows);
-    final tempFilePath = tempPath.resolve(fileName!).path;
+    final tempFilePath = tempPath
+        .resolve(fileName!)
+        .toFilePath(windows: CurrentPlatform.isWindows);
     print(tempFilePath);
 
     final attachmentFileBytes = attachmentFile.bytes;
