@@ -17,6 +17,7 @@ class StreamVideoAttachment extends StreamAttachmentWidget {
     this.onShowMessage,
     this.onReplyMessage,
     this.onAttachmentTap,
+    this.attachmentActionsModalBuilder,
   });
 
   /// The [StreamMessageThemeData] to use for the title
@@ -31,6 +32,9 @@ class StreamVideoAttachment extends StreamAttachmentWidget {
   /// {@macro onAttachmentTap}
   final OnAttachmentTap? onAttachmentTap;
 
+  /// {@macro attachmentActionsBuilder}
+  final AttachmentActionsBuilder? attachmentActionsModalBuilder;
+
   @override
   Widget build(BuildContext context) {
     return source.when(
@@ -41,7 +45,7 @@ class StreamVideoAttachment extends StreamAttachmentWidget {
         return _buildVideoAttachment(
           context,
           StreamVideoThumbnailImage(
-            video: attachment.file!.path!,
+            video: attachment.file!.path,
             thumbUrl: attachment.thumbUrl,
             constraints: constraints,
           ),
@@ -54,7 +58,7 @@ class StreamVideoAttachment extends StreamAttachmentWidget {
         return _buildVideoAttachment(
           context,
           StreamVideoThumbnailImage(
-            video: attachment.assetUrl!,
+            video: attachment.assetUrl,
             thumbUrl: attachment.thumbUrl,
             constraints: constraints,
           ),
@@ -86,6 +90,8 @@ class StreamVideoAttachment extends StreamAttachmentWidget {
                               userName: message.user!.name,
                               onShowMessage: onShowMessage,
                               onReplyMessage: onReplyMessage,
+                              attachmentActionsModalBuilder:
+                                  attachmentActionsModalBuilder,
                             ),
                           ),
                         ),
