@@ -18,6 +18,9 @@ abstract class Translations {
   /// The label for showing no users
   String get noUsersLabel;
 
+  /// The label for showing no photo or video
+  String get noPhotoOrVideoLabel;
+
   /// The text for showing user is online
   String get userOnlineText;
 
@@ -78,7 +81,10 @@ abstract class Translations {
 
   /// The text for showing the unread messages count
   /// in the [StreamMessageListView]
-  String unreadMessagesSeparatorText(int unreadCount);
+  String unreadMessagesSeparatorText(
+    @Deprecated('unreadCount is not used anymore and will be removed ')
+        int unreadCount,
+  );
 
   /// The label for "connected" in [StreamConnectionStatusBuilder]
   String get connectedLabel;
@@ -372,6 +378,9 @@ class DefaultTranslations implements Translations {
 
   @override
   String get noUsersLabel => 'There are no users currently';
+
+  @override
+  String get noPhotoOrVideoLabel => 'There is no photo or video';
 
   @override
   String get retryLabel => 'Retry';
@@ -791,12 +800,7 @@ Attachment limit exceeded: it's not possible to add more than $limit attachments
   String get linkDisabledError => 'Links are disabled';
 
   @override
-  String unreadMessagesSeparatorText(int unreadCount) {
-    if (unreadCount == 1) {
-      return '1 unread message';
-    }
-    return '$unreadCount unread messages';
-  }
+  String unreadMessagesSeparatorText(int unreadCount) => 'New messages';
 
   @override
   String get enableFileAccessMessage => 'Please enable access to files'
