@@ -276,7 +276,7 @@ class _ParseAttachments extends StatelessWidget {
       'video': (_, attachment) {
         return StreamVideoThumbnailImage(
           key: ValueKey(attachment.assetUrl),
-          video: attachment.file?.path ?? attachment.assetUrl!,
+          video: attachment.file?.path ?? attachment.assetUrl,
           constraints: BoxConstraints.loose(const Size(32, 32)),
           errorBuilder: (_, __) => AttachmentError(
             constraints: BoxConstraints.loose(const Size(32, 32)),
