@@ -40,7 +40,8 @@ class StreamMessageText extends StatelessWidget {
             .translate(language)
             .replaceMentions()
             .text
-            ?.replaceAll('\n', '\n\n');
+            ?.replaceAll('\n', '\n\n')
+            .trim();
         final themeData = Theme.of(context);
         return MarkdownBody(
           data: messageText ?? '',
