@@ -8,6 +8,7 @@ import 'package:stream_chat_flutter/src/context_menu_items/context_menu_reaction
 import 'package:stream_chat_flutter/src/context_menu_items/stream_chat_context_menu_item.dart';
 import 'package:stream_chat_flutter/src/dialogs/dialogs.dart';
 import 'package:stream_chat_flutter/src/message_actions_modal/message_actions_modal.dart';
+import 'package:stream_chat_flutter/src/message_widget/bottom_row.dart';
 import 'package:stream_chat_flutter/src/message_widget/message_widget_content.dart';
 import 'package:stream_chat_flutter/src/message_widget/reactions/message_reactions_modal.dart';
 import 'package:stream_chat_flutter/stream_chat_flutter.dart';
@@ -79,8 +80,15 @@ class StreamMessageWidget extends StatefulWidget {
     this.userAvatarBuilder,
     this.editMessageInputBuilder,
     this.textBuilder,
-    this.bottomRowBuilder,
-    this.deletedBottomRowBuilder,
+    @Deprecated('''
+    Use [bottomRowBuilderWithDefaultWidget] instead.
+    Will be removed in the next major version.
+    ''') this.bottomRowBuilder,
+    this.bottomRowBuilderWithDefaultWidget,
+    @Deprecated('''
+    Use [bottomRowBuilderWithDefaultWidget] instead.
+    Will be removed in the next major version.
+    ''') this.deletedBottomRowBuilder,
     this.customAttachmentBuilders,
     this.padding,
     this.textPadding = const EdgeInsets.symmetric(
@@ -92,11 +100,19 @@ class StreamMessageWidget extends StatefulWidget {
     this.onQuotedMessageTap,
     this.customActions = const [],
     this.onAttachmentTap,
-    this.usernameBuilder,
+    @Deprecated('''
+    Use [bottomRowBuilderWithDefaultWidget] instead.
+    Will be removed in the next major version.
+    ''') this.usernameBuilder,
     this.imageAttachmentThumbnailSize = const Size(400, 400),
     this.imageAttachmentThumbnailResizeType = 'clip',
     this.imageAttachmentThumbnailCropType = 'center',
-  }) : attachmentBuilders = {
+    this.attachmentActionsModalBuilder,
+  })  : assert(
+          bottomRowBuilder == null || bottomRowBuilderWithDefaultWidget == null,
+          'You can only use one of the two bottom row builders',
+        ),
+        attachmentBuilders = {
           'image': (context, message, attachments) {
             final border = RoundedRectangleBorder(
               side: attachmentBorderSide ??
@@ -129,6 +145,8 @@ class StreamMessageWidget extends StatefulWidget {
                       imageThumbnailResizeType:
                           imageAttachmentThumbnailResizeType,
                       imageThumbnailCropType: imageAttachmentThumbnailCropType,
+                      attachmentActionsModalBuilder:
+                          attachmentActionsModalBuilder,
                     ),
                   ),
                   attachmentShape: border,
@@ -156,6 +174,7 @@ class StreamMessageWidget extends StatefulWidget {
                 imageThumbnailSize: imageAttachmentThumbnailSize,
                 imageThumbnailResizeType: imageAttachmentThumbnailResizeType,
                 imageThumbnailCropType: imageAttachmentThumbnailCropType,
+                attachmentActionsModalBuilder: attachmentActionsModalBuilder,
               ),
               attachmentShape: border,
             );
@@ -189,6 +208,8 @@ class StreamMessageWidget extends StatefulWidget {
                             onAttachmentTap(message, attachment);
                           }
                         : null,
+                    attachmentActionsModalBuilder:
+                        attachmentActionsModalBuilder,
                   );
                 }).toList(),
               ),
@@ -306,7 +327,13 @@ class StreamMessageWidget extends StatefulWidget {
   /// {@template bottomRowBuilder}
   /// Widget builder for building a bottom row below the message
   /// {@endtemplate}
-  final Widget Function(BuildContext, Message)? bottomRowBuilder;
+  final BottomRowBuilder? bottomRowBuilder;
+
+  /// {@template bottomRowBuilderWithDefaultWidget}
+  /// Widget builder for building a bottom row below the message.
+  /// Also contains the default bottom row widget.
+  /// {@endtemplate}
+  final BottomRowBuilderWithDefaultWidget? bottomRowBuilderWithDefaultWidget;
 
   /// {@template deletedBottomRowBuilder}
   /// Widget builder for building a bottom row below a deleted message
@@ -512,6 +539,9 @@ class StreamMessageWidget extends StatefulWidget {
   /// {@macro onMessageWidgetAttachmentTap}
   final OnMessageWidgetAttachmentTap? onAttachmentTap;
 
+  /// {@macro attachmentActionsBuilder}
+  final AttachmentActionsBuilder? attachmentActionsModalBuilder;
+
   /// Size of the image attachment thumbnail.
   final Size imageAttachmentThumbnailSize;
 
@@ -537,9 +567,19 @@ class StreamMessageWidget extends StatefulWidget {
     void Function(Message)? onReplyTap,
     Widget Function(BuildContext, Message)? editMessageInputBuilder,
     Widget Function(BuildContext, Message)? textBuilder,
-    Widget Function(BuildContext, Message)? usernameBuilder,
-    Widget Function(BuildContext, Message)? bottomRowBuilder,
-    Widget Function(BuildContext, Message)? deletedBottomRowBuilder,
+    @Deprecated('''
+    Use [bottomRowBuilderWithDefaultWidget] instead.
+    Will be removed in the next major version.
+    ''') Widget Function(BuildContext, Message)? usernameBuilder,
+    @Deprecated('''
+    Use [bottomRowBuilderWithDefaultWidget] instead.
+    Will be removed in the next major version.
+    ''') BottomRowBuilder? bottomRowBuilder,
+    BottomRowBuilderWithDefaultWidget? bottomRowBuilderWithDefaultWidget,
+    @Deprecated('''
+    Use [bottomRowBuilderWithDefaultWidget] instead.
+    Will be removed in the next major version.
+    ''') Widget Function(BuildContext, Message)? deletedBottomRowBuilder,
     void Function(BuildContext, Message)? onMessageActions,
     Message? message,
     StreamMessageThemeData? messageTheme,
@@ -586,7 +626,31 @@ class StreamMessageWidget extends StatefulWidget {
     Size? imageAttachmentThumbnailSize,
     String? imageAttachmentThumbnailResizeType,
     String? imageAttachmentThumbnailCropType,
+    AttachmentActionsBuilder? attachmentActionsModalBuilder,
   }) {
+    assert(
+      bottomRowBuilder == null || bottomRowBuilderWithDefaultWidget == null,
+      'You can only use one of the two bottom row builders',
+    );
+
+    var _bottomRowBuilderWithDefaultWidget =
+        bottomRowBuilderWithDefaultWidget ??
+            this.bottomRowBuilderWithDefaultWidget;
+
+    _bottomRowBuilderWithDefaultWidget ??= (context, message, defaultWidget) {
+      final _bottomRowBuilder = bottomRowBuilder ?? this.bottomRowBuilder;
+      if (_bottomRowBuilder != null) {
+        return _bottomRowBuilder(context, message);
+      }
+
+      return defaultWidget.copyWith(
+        onThreadTap: onThreadTap ?? this.onThreadTap,
+        usernameBuilder: usernameBuilder ?? this.usernameBuilder,
+        deletedBottomRowBuilder:
+            deletedBottomRowBuilder ?? this.deletedBottomRowBuilder,
+      );
+    };
+
     return StreamMessageWidget(
       key: key ?? this.key,
       onMentionTap: onMentionTap ?? this.onMentionTap,
@@ -595,10 +659,7 @@ class StreamMessageWidget extends StatefulWidget {
       editMessageInputBuilder:
           editMessageInputBuilder ?? this.editMessageInputBuilder,
       textBuilder: textBuilder ?? this.textBuilder,
-      usernameBuilder: usernameBuilder ?? this.usernameBuilder,
-      bottomRowBuilder: bottomRowBuilder ?? this.bottomRowBuilder,
-      deletedBottomRowBuilder:
-          deletedBottomRowBuilder ?? this.deletedBottomRowBuilder,
+      bottomRowBuilderWithDefaultWidget: _bottomRowBuilderWithDefaultWidget,
       onMessageActions: onMessageActions ?? this.onMessageActions,
       message: message ?? this.message,
       messageTheme: messageTheme ?? this.messageTheme,
@@ -652,6 +713,8 @@ class StreamMessageWidget extends StatefulWidget {
           this.imageAttachmentThumbnailResizeType,
       imageAttachmentThumbnailCropType: imageAttachmentThumbnailCropType ??
           this.imageAttachmentThumbnailCropType,
+      attachmentActionsModalBuilder:
+          attachmentActionsModalBuilder ?? this.attachmentActionsModalBuilder,
     );
   }
 
@@ -838,52 +901,69 @@ class _StreamMessageWidgetState extends State<StreamMessageWidget>
                       ? Alignment.centerRight
                       : Alignment.centerLeft,
                   widthFactor: widget.widthFactor,
-                  child: MessageWidgetContent(
-                    streamChatTheme: _streamChatTheme,
-                    showUsername: showUsername,
-                    showTimeStamp: showTimeStamp,
-                    showThreadReplyIndicator: showThreadReplyIndicator,
-                    showSendingIndicator: showSendingIndicator,
-                    showInChannel: showInChannel,
-                    isGiphy: isGiphy,
-                    isOnlyEmoji: isOnlyEmoji,
-                    hasUrlAttachments: hasUrlAttachments,
-                    messageTheme: widget.messageTheme,
-                    reverse: widget.reverse,
-                    message: widget.message,
-                    hasNonUrlAttachments: hasNonUrlAttachments,
-                    shouldShowReactions: shouldShowReactions,
-                    hasQuotedMessage: hasQuotedMessage,
-                    textPadding: widget.textPadding,
-                    attachmentBuilders: widget.attachmentBuilders,
-                    attachmentPadding: widget.attachmentPadding,
-                    avatarWidth: avatarWidth,
-                    bottomRowPadding: bottomRowPadding,
-                    isFailedState: isFailedState,
-                    isPinned: isPinned,
-                    messageWidget: widget,
-                    showBottomRow: showBottomRow,
-                    showPinHighlight: widget.showPinHighlight,
-                    showReactionPickerIndicator:
-                        widget.showReactionPickerIndicator,
-                    showReactions: showReactions,
-                    showUserAvatar: widget.showUserAvatar,
-                    streamChat: _streamChat,
-                    translateUserAvatar: widget.translateUserAvatar,
-                    deletedBottomRowBuilder: widget.deletedBottomRowBuilder,
-                    onThreadTap: widget.onThreadTap,
-                    shape: widget.shape,
-                    borderSide: widget.borderSide,
-                    borderRadiusGeometry: widget.borderRadiusGeometry,
-                    textBuilder: widget.textBuilder,
-                    onLinkTap: widget.onLinkTap,
-                    onMentionTap: widget.onMentionTap,
-                    onQuotedMessageTap: widget.onQuotedMessageTap,
-                    bottomRowBuilder: widget.bottomRowBuilder,
-                    onUserAvatarTap: widget.onUserAvatarTap,
-                    userAvatarBuilder: widget.userAvatarBuilder,
-                    usernameBuilder: widget.usernameBuilder,
-                  ),
+                  child: Builder(builder: (context) {
+                    var _bottomRowBuilderWithDefaultWidget =
+                        widget.bottomRowBuilderWithDefaultWidget;
+
+                    _bottomRowBuilderWithDefaultWidget ??=
+                        (context, message, defaultWidget) {
+                      final _bottomRowBuilder = widget.bottomRowBuilder;
+                      if (_bottomRowBuilder != null) {
+                        return _bottomRowBuilder(context, message);
+                      }
+
+                      return defaultWidget.copyWith(
+                        onThreadTap: widget.onThreadTap,
+                        usernameBuilder: widget.usernameBuilder,
+                        deletedBottomRowBuilder: widget.deletedBottomRowBuilder,
+                      );
+                    };
+
+                    return MessageWidgetContent(
+                      streamChatTheme: _streamChatTheme,
+                      showUsername: showUsername,
+                      showTimeStamp: showTimeStamp,
+                      showThreadReplyIndicator: showThreadReplyIndicator,
+                      showSendingIndicator: showSendingIndicator,
+                      showInChannel: showInChannel,
+                      isGiphy: isGiphy,
+                      isOnlyEmoji: isOnlyEmoji,
+                      hasUrlAttachments: hasUrlAttachments,
+                      messageTheme: widget.messageTheme,
+                      reverse: widget.reverse,
+                      message: widget.message,
+                      hasNonUrlAttachments: hasNonUrlAttachments,
+                      shouldShowReactions: shouldShowReactions,
+                      hasQuotedMessage: hasQuotedMessage,
+                      textPadding: widget.textPadding,
+                      attachmentBuilders: widget.attachmentBuilders,
+                      attachmentPadding: widget.attachmentPadding,
+                      avatarWidth: avatarWidth,
+                      bottomRowPadding: bottomRowPadding,
+                      isFailedState: isFailedState,
+                      isPinned: isPinned,
+                      messageWidget: widget,
+                      showBottomRow: showBottomRow,
+                      showPinHighlight: widget.showPinHighlight,
+                      showReactionPickerIndicator:
+                          widget.showReactionPickerIndicator,
+                      showReactions: showReactions,
+                      showUserAvatar: widget.showUserAvatar,
+                      streamChat: _streamChat,
+                      translateUserAvatar: widget.translateUserAvatar,
+                      shape: widget.shape,
+                      borderSide: widget.borderSide,
+                      borderRadiusGeometry: widget.borderRadiusGeometry,
+                      textBuilder: widget.textBuilder,
+                      onLinkTap: widget.onLinkTap,
+                      onMentionTap: widget.onMentionTap,
+                      onQuotedMessageTap: widget.onQuotedMessageTap,
+                      bottomRowBuilderWithDefaultWidget:
+                          _bottomRowBuilderWithDefaultWidget,
+                      onUserAvatarTap: widget.onUserAvatarTap,
+                      userAvatarBuilder: widget.userAvatarBuilder,
+                    );
+                  }),
                 ),
               ),
             ),
