@@ -92,9 +92,9 @@ void main() {
 
       const errorMessage = 'Error! Error! Error!';
       final error = DioError(
-        type: DioErrorType.response,
-        error: errorMessage,
-        requestOptions: RequestOptions(path: ''),
+        type: DioErrorType.badResponse,
+        message: errorMessage,
+        requestOptions: RequestOptions(),
       );
       when(() => mockChannel.initialized)
           .thenAnswer((_) => Future.error(error));
