@@ -15,6 +15,9 @@ class StreamChatLocalizationsIt extends GlobalStreamChatLocalizations {
   String get noUsersLabel => "Non c'é nessun utente al momento";
 
   @override
+  String get noPhotoOrVideoLabel => 'Non ci sono foto o video';
+
+  @override
   String get retryLabel => 'Riprova';
 
   @override
@@ -439,12 +442,7 @@ Attenzione: il limite massimo di $limit file è stato superato.
   String get linkDisabledError => 'I links sono disattivati';
 
   @override
-  String unreadMessagesSeparatorText(int unreadCount) {
-    if (unreadCount == 1) {
-      return '1 messaggio non letto';
-    }
-    return '$unreadCount messaggi non letti';
-  }
+  String unreadMessagesSeparatorText(int unreadCount) => 'Nouveaux messages';
 
   @override
   String get enableFileAccessMessage => "Per favore attiva l'accesso ai file"
